refactor of package and fix tests
reorganising
