+++ molecularnodes/io/ensemble/cellpack.py
@@ -6,7 +6,7 @@ import bpy
 from .ensemble import Ensemble
 from .bcif import BCIF
 from .cif import OldCIF
-from ..parse import molecule
+from ..molecule import molecule
 from ... import blender as bl
 from ... import color
