+++ tests/test_star.py
@@ -10,7 +10,7 @@ mn._test_register()
 @pytest.mark.parametrize("type", ["cistem", "relion"])
 def test_starfile_attributes(type):
     file = data_dir / f"{type}.star"
-    ensemble = mn.io.star.load(file)
+    ensemble = mn.io.ensemble.load_starfile(file)
 
     star = starfile.read(file)
 
@@ -43,15 +43,13 @@ def test_starfile_attributes(type):
 
 def test_categorical_attributes():
     file = data_dir / "cistem.star"
-    ensemble = mn.io.star.load(file)
+    ensemble = mn.io.ensemble.load_starfile(file)
     assert "cisTEMOriginalImageFilename_categories" in ensemble.object
 
 
 def test_micrograph_conversion():
-    from pathlib import Path
-
     file = data_dir / "cistem.star"
-    ensemble = mn.io.star.load(file)
+    ensemble = mn.io.ensemble.load_starfile(file)
     tiff_path = data_dir / "montage.tiff"
     tiff_path.unlink(missing_ok=True)
     ensemble._convert_mrc_to_tiff()
@@ -65,7 +63,7 @@ def test_micrograph_loading():
     tiff_path = data_dir / "montage.tiff"
     tiff_path.unlink(missing_ok=True)
 
-    ensemble = mn.io.star.load(file)
+    ensemble = mn.io.ensemble.load_starfile(file)
     assert not tiff_path.exists()
     ensemble.star_node.inputs["Show Micrograph"].default_value = True
     bpy.context.evaluated_depsgraph_get().update()
